streaming/requirements.txt
# Streaming Dependencies
kafka-python==2.0.2
confluent-kafka==2.2.0
# Для симуляции данных
numpy==1.24.3
scipy==1.11.4
# Для работы с данными
pandas==2.1.4
# Для мониторинга
prometheus-client==0.19.0
# Для конфигурации
pyyaml==6.0.1
python-dotenv==1.0.0